src/test/java/com/example/springboot/controller/UserControllerTest.java
package com.example.springboot.controller;

import com.example.springboot.model.User;
import com.example.springboot.service.UserService;
import com.example.springboot.utils.Constants;
import com.example.springboot.utils.ObjectMapperUtils;
import com.fasterxml.jackson.databind.ObjectMapper;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.MockMvcBuilder;
import org.springframework.test.web.servlet.MvcResult;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertNotNull;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.when;

@RunWith(SpringRunner.class)
@WebMvcTest(UserController.class)
public class UserControllerTest {
    @Autowired
    private MockMvc mvc;

    @MockBean
    private UserService userService;

    @Test
    public void testCreateUser() throws Exception {
        User user = new User("Kaushik", Constants.Gender.MALE);
        String inputJson = ObjectMapperUtils.mapToJson(user);

        when(userService.createUser(any())).thenReturn(user);
        MvcResult mvcResult = mvc.perform(MockMvcRequestBuilders.post("/users")
                .contentType(MediaType.APPLICATION_JSON_VALUE).content(inputJson)).andReturn();
        assertNotNull(mvcResult);
        String response = mvcResult.getResponse().getContentAsString();
        User responseUser = ObjectMapperUtils.mapFromJson(response, User.class);
        assertNotNull(responseUser);
        assertEquals(responseUser.getUserName(), user.getUserName());
    }
}
